Unified instance internal API
