+++ nginx/conf.d/app.conf
@@ -4,6 +4,10 @@ server {
   server_name ~.;
 
   location /graphql {
+    auth_request /_apikey_introspection;
+    auth_request_set $dgraph_token $upstream_http_x_dgraph_accesstoken;
+    proxy_set_header	X-Dgraph-AccessToken		$dgraph_token;
+
     proxy_set_header	Host		$host;
     proxy_set_header	X-Real_IP	$remote_addr;
     proxy_pass http://alpha:8080/graphql;
@@ -11,4 +15,18 @@ server {
 
   root /media/data;
   autoindex on;
+
+  location = /_apikey_introspection {
+    internal;
+    proxy_method      POST;
+    proxy_set_header	Host		$host;
+    proxy_set_header	X-Real_IP	$remote_addr;
+    proxy_pass        http://app_server:5000/apikey;
+
+    proxy_cache           apikey_responses;         # Enable caching
+    proxy_cache_key       $http_x_api_key;          # Cache for each access token
+    proxy_cache_lock      on;                       # Duplicate tokens must wait
+    proxy_cache_valid     200 5s;                   # How long to use each response
+    proxy_ignore_headers  Cache-Control Expires Set-Cookie;
+  }
 }
