+++ nginx/nginx.conf
@@ -9,5 +9,6 @@ events {
 }
 
 http {
+  proxy_cache_path /var/cache/nginx/apikey keys_zone=apikey_responses:1m max_size=2m;
   include /etc/nginx/conf.d/app.conf;
 }
